Refactor step 1: replace double with float
- torque_control(): static double time -> static float time
  Also fix 1/1000.0 integer division to explicit 0.001f literal
- hes_t.counter: double -> float

The Cortex-M4 FPU only handles single-precision. double falls back
to soft-float (~10x slower). No precision loss for these use cases.

+++ Core/Inc/FOC.h
@@ -193,7 +193,7 @@ typedef struct {
 	int int_angle;
 	SPI_HandleTypeDef *spiHandle;
 	uint8_t HES_RxBuf[2];
-	double counter;
+	float counter;
 
 } hes_t;
